pyproject.toml
[tool.poetry]
name = "ludoteca"
version = "0.1.0"
description = ""
authors = ["Fábio Ferreira <fabiogasparferreira@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.8"
django = "^3.2.8"
djangorestframework = "^3.12.4"
django-filter = "^21.1"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.0"
boardgamegeek2 = "^1.0.1"
pandas = "^1.3.3"
djangorestframework-simplejwt = "^5.2.2"
PyJWT= "2.1.0"
django-crispy-forms = "^1.13.0"
phonenumberslite = "^8.12.34"
django-phonenumber-field = "^5.2.0"
django-rest-passwordreset = "^1.3.0"
requests-cache = "0.5.2"
drf-yasg = "^1.20.0"
Pillow = "^9.2.0"
django-environ = "^0.9.0"
python-dotenv = "^0.20.0"
django-polymorphic = "^3.1.0"
django-rest-polymorphic = "^0.1.10"
qrcode = "^7.4.2"
django-background-tasks = "^1.2.5"

[tool.poetry.dev-dependencies]
pylint = "^2.11.1"
black = "^21.9b0"
isort = "^5.9.3"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
